fix(collections): modify `Rank`
Modify `Rank` type of all collection classes to `size_t` by type-alias, and change its scope from namespace `dsa` to single file.

+++ test/test_vector.cpp
@@ -7,7 +7,7 @@ TEST_CASE("Vector initialize", "[Vector]") {
     dsa::Vector<int> v2(v1);
     // TODO: `==` operator is not implemented
     REQUIRE(v1.size() == v2.size());
-    for (int i = 0; i < v1.size(); ++i) {
+    for (size_t i = 0; i < v1.size(); ++i) {
         REQUIRE(v1[i] == v2[i]);
     }
 }
